backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
httpx
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
supabase
python-dotenv
resend
alembic
